% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dists.R
\name{get_dists}
\alias{get_dists}
\title{Retrieve distances of an elment i,j from a 'dist' object.}
\usage{
get_dists(x, i, j = NULL, parallel = TRUE)
}
\arguments{
\item{x}{An object of class 'dist'.}

\item{i}{Either a two column matrix of integer indices, a vector of indices
that are paired with the corresponding indices in j.}

\item{j}{A vector of indices where each element forms a pair with the
corresponding element in argument i.}

\item{parallel}{Logical indicating whether the operation should be
parallelized.}
}
\value{
A vector giving the distances between the provided indices.
}
\description{
Retrieve distances of an elment i,j from a 'dist' object.
}
\examples{
test <- matrix(c(rnorm(10), rnorm(10)-5, rnorm(10)+5), ncol = 2)
test_dists <- dist(test)
indices <- matrix(sample(1:8),ncol=2)
get_dists(test_dists, indices)
}
